+++ src/index.js
@@ -4,6 +4,35 @@ const Semver = require('semver');
 const { parser, Release, Changelog } = require('keep-a-changelog');
 const fs = require('fs');
 const chalk = require('chalk');
+const url = require('url');
+
+function getPackageUrl (packageUrl) {
+  packageUrl = packageUrl
+    //trim git+ from the front of the URL
+    .replace(/^git\+/, '')
+    //convert ssh://, ssh://git@, or just git@ to https://
+    .replace(/^(?:ssh:\/\/(?:git@)?|git@)/, 'https://');
+
+  const parsed = url.parse(packageUrl);
+  
+  if (parsed.pathname) {
+    parsed.pathname = parsed.pathname
+      //replace the leading : for ssh paths with / (and pull the port out of the parsed path)
+      .replace(/\/:(?:(\d+):)?/, (match, port) => {
+        if (!parsed.port && port) {
+          parsed.port = port;
+          parsed.host += ':' + port;
+        }
+        return '/';
+      })
+      //replace the trailing .git
+      .replace(/\.git$/, '');
+  }
+
+  return url.format(parsed);
+}
+
+module.exports.getPackageUrl = getPackageUrl;
 
 function getPackage () {
   let pkg;
@@ -17,8 +46,10 @@ function getPackage () {
     throw new Error('No repository URL found in package.json');
   }
 
+  pkg.repository.url = getPackageUrl(pkg.repository.url);
+
   if (!/^https?:\/\//.test(pkg.repository.url)) {
-    throw new Error('Repository URL in package.json must be an http or https URL');
+    throw new Error('Repository URL in package.json must be a valid URL to a git repository');
   }
   return pkg;
 }